requirements.txt
boto3
faiss-cpu
sentence-transformers
numpy